requirements.txt
altgraph==0.17.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
beautifulsoup4==4.14.2
billiard==4.2.2
blinker==1.9.0
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
distro==1.9.0
dj-database-url==3.1.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
faiss-cpu==1.12.0
filelock==3.20.0
Flask==3.1.2
fonttools==4.61.1
fsspec==2025.10.0
gunicorn==24.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
kiwisolver==1.4.9
kombu==5.5.4
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.5
nltk==3.9.2
numpy==2.3.4
openai==2.7.2
opencv-python==4.13.0.90
packaging==25.0
pdfminer.six==20251107
pdfplumber==0.11.8
pefile==2023.2.7
pillow==12.0.0
polars==1.37.1
polars-runtime-32==1.37.1
prompt_toolkit==3.0.52
psutil==7.2.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyMuPDF==1.26.6
pyparsing==3.3.2
PyPDF2==3.0.1
pypdfium2==5.0.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
pywin32-ctypes==0.2.3
PyYAML==6.0.3
redis==7.0.1
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.4.7
ultralytics-thop==2.0.18
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.5
whitenoise==6.11.0
